requirements.txt
anthropic>=0.18.0
pydantic>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
requests>=2.28.0
celery>=5.3.0
redis>=5.0.0
eventlet>=0.33.0
pypdf>=3.0.0
Pillow>=10.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
